Moved mute and selfmute commands to respective cogs
